Allow dnf-json to be mocked
We want to test API methods which calls dnf. Unfortunately, calling dnf
is expensive operation - it requires network access and downloading
a lot of (meta)data. This commit changes the former rpmmd implementation
so that it can be mocked.

This commit also changes the name of rpmmd package to dnfadapter as I
think it better represents what it actually does.