test/cases: update for rpm move
Given the move from dnf to rpm the test cases need to use the rpm stage
in the pipeline build and in the stages. The base packages, build
packages, excluded packages, module platform id, arch, release version,
and repos are no longer explicitely needed by the pipeline. The base
and excluded packages are exposed by the distro configs for each image
type. The build packages are exposed by the distro configs for each
distro. The ModulePlatformID is also exposed by the distro configs.