tools: add script to generate multiple test cases
This script generates json test cases for each supported output format.
It requires an architecure, distro, osbuild store, and output directory
as input. It formats and passes this info to the generate-test-case
script which creates the individual test cases. Then this script writes
these json test cases to files in the specified output directory.