rpmmd: add ssl fields to RepoConfig
In order to connect to RHEL's CDN dnf must pass an SSL CA Cert, an SSL
Client Key, and an SSL Client cert. The RepoConfig now contains these
additional fields. When LoadRepositories is called, if the distro is
RHEL 8, the needed cert and key files will be added to the RepoConfig
which gets passed to dnf-json.